Move world gen into separate module, add berries
